added bytes

+++ libs/mcp-core/src/models/types.rs
@@ -293,7 +293,7 @@ pub struct RegistryTool {
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct RegistryToolsResponse {
     pub count: u32,
-    pub version: u32,
+    pub version: u64,
     pub categories: HashMap<String, u32>,
     pub tags: HashMap<String, u32>,
     pub tools: Vec<RegistryTool>,
